+++ djpcms/contrib/flowrepo/tests/__init__.py
@@ -1,2 +1,3 @@
 from forms import *
 from cms import *
+from markups import *